Land PR 266, but with corrected/guessed metadata
Original PR branch was deleted, current content looks impossible
to be the actual original content of the PR, in terms of metadata.

The author metadata for this commit is corrected/fabricated/guessed,
based on PR submitter/date, so as to be something much more plausible.

Also, turned it into a regular, basic commit, rebased on master,
instead of it being a merge commit of another PR... (What the heck????)